Restructure state and reducers
