+++ applications/dashboard/src/scripts/app/index.tsx
@@ -8,30 +8,30 @@
 import { addRoutes, addComponent, onReady } from "@dashboard/application";
 import React from "react";
 import { Route } from "react-router-dom";
-import SignInPage from "./authenticate/SignInPage";
-import PasswordPage from "./authenticate/PasswordPage";
-import RecoverPasswordPage from "./authenticate/RecoverPasswordPage";
+import SignInPage from "@dashboard/app/authenticate/SignInPage";
+import PasswordPage from "@dashboard/app/authenticate/PasswordPage";
+import RecoverPasswordPage from "@dashboard/app/authenticate/RecoverPasswordPage";
 import Router from "@dashboard/components/Router";
 import { stickyHeader } from "@dashboard/dom";
 
 // These imports are all responsible for initializing themselves.
-import "./user-content/emoji";
-import "./user-content/spoilers";
-import "./user-content/embeds/image";
-import "./user-content/embeds/link";
-import "./user-content/embeds/twitter";
-import "./user-content/embeds/video";
-import "./user-content/embeds/instagram";
-import "./user-content/embeds/imgur";
-import "./user-content/embeds/soundcloud";
-import "./user-content/embeds/getty";
-import "./user-content/embeds/giphy";
-import "./user-content/embeds/codepen";
+import "@dashboard/app/user-content/emoji";
+import "@dashboard/app/user-content/spoilers";
+import "@dashboard/app/user-content/embeds/image";
+import "@dashboard/app/user-content/embeds/link";
+import "@dashboard/app/user-content/embeds/twitter";
+import "@dashboard/app/user-content/embeds/video";
+import "@dashboard/app/user-content/embeds/instagram";
+import "@dashboard/app/user-content/embeds/imgur";
+import "@dashboard/app/user-content/embeds/soundcloud";
+import "@dashboard/app/user-content/embeds/getty";
+import "@dashboard/app/user-content/embeds/giphy";
+import "@dashboard/app/user-content/embeds/codepen";
 import { registerReducer } from "@dashboard/state/reducerRegistry";
-import authenticateReducer from "@dashboard/state/authenticate/authenticateReducer";
+import sessionReducer from "@dashboard/state/session/sessionReducer";
 
 onReady(() => {
-    registerReducer("authenticate", authenticateReducer);
+    registerReducer("session", sessionReducer);
     addComponent("App", Router);
 
     addRoutes([
