+++ applications/dashboard/src/scripts/components/PageLoading.tsx
@@ -1,7 +1,6 @@
 import * as React from "react";
-import { ILoadable } from "@dashboard/state/IState";
-import { LoadStatus } from "@dashboard/state/IState";
 import { t } from "@dashboard/application";
+import { ILoadable, LoadStatus } from "@dashboard/types/api";
 
 export default class PageLoading extends React.PureComponent<ILoadable<any>, {}> {
     public static defaultProps: Partial<ILoadable<any>> = {
