+++ applications/dashboard/src/scripts/state/session/authenticatorsActions.ts
@@ -5,12 +5,11 @@
 
 import { Dispatch } from "redux";
 import { generateApiActionCreators, ActionsUnion, createAction } from "@dashboard/state/utility";
-import api, { LoadStatus } from "@dashboard/apiv2";
-import IState from "@dashboard/state/IState";
-import { IAuthenticator } from "@dashboard/state/authenticate/IAuthenticateState";
+import api from "@dashboard/apiv2";
 import { AxiosResponse } from "axios";
+import { LoadStatus, IUserAuthenticator } from "@dashboard/types/api";
+import { IStoreState } from "@dashboard/types/state";
 
-// export const GET_SIGNIN_AUTHENTICATORS = 'GET_SIGNIN_AUTHENTICATORS';
 export const GET_SIGNIN_AUTHENTICATORS_REQUEST = "GET_SIGNIN_AUTHENTICATORS_REQUEST";
 export const GET_SIGNIN_AUTHENTICATORS_ERROR = "GET_SIGNIN_AUTHENTICATORS_ERROR";
 export const GET_SIGNIN_AUTHENTICATORS_SUCCESS = "GET_SIGNIN_AUTHENTICATORS_SUCCESS";
@@ -20,21 +19,16 @@ const getAuthenticatorsActions = generateApiActionCreators(
     GET_SIGNIN_AUTHENTICATORS_SUCCESS,
     GET_SIGNIN_AUTHENTICATORS_ERROR,
     // https://github.com/Microsoft/TypeScript/issues/10571#issuecomment-345402872
-    [] as IAuthenticator[],
+    [] as IUserAuthenticator[],
 );
 
 export function getAuthenticators() {
-    return (dispatch: Dispatch<any>, getState: () => IState) => {
-        const { authenticate } = getState();
-        if (authenticate.signin.status === LoadStatus.LOADING) {
-            return;
-        }
-
+    return (dispatch: Dispatch<any>, getState: () => IStoreState) => {
         dispatch(getAuthenticatorsActions.request());
 
         return api
             .get("/authenticate/authenticators")
-            .then((response: AxiosResponse<IAuthenticator[]>) => {
+            .then((response: AxiosResponse<IUserAuthenticator[]>) => {
                 dispatch(getAuthenticatorsActions.success(response));
             })
             .catch(error => {
