+++ applications/dashboard/src/scripts/state/session/sessionReducer.ts
@@ -1,0 +1,14 @@
+/**
+ * @author Adam (charrondev) Charron <adam.c@vanillaforums.com>
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license https://opensource.org/licenses/GPL-2.0 GPL-2.0
+ */
+
+import { combineReducers } from "redux";
+import authenticatorsReducer from "@dashboard/state/session/authenticatorsReducer";
+
+const sessionReducer = combineReducers({
+    authenticator: authenticatorsReducer,
+});
+
+export default sessionReducer;
