+++ applications/dashboard/src/scripts/types/global.d.ts
