+++ applications/dashboard/src/scripts/types/store/IAuthenticateState.ts
@@ -1,0 +1,4 @@
+/**
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license http://www.opensource.org/licenses/gpl-2.0.php GNU GPL v2
+ */
